Moved all methods from NodeWithConstructors to NodeWithMembers and removed NodeWithConstructors completely.
